# public/doc_swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'Api oAuth 2.0'
  description: 'Esta é a documentação base da sua API. Ela contém todas as informações necessárias para que você, sua equipe ou um desenvolvedor possam integrar o seu sistema com outros sistemas, sites, portais ou aplicativos móveis (iOS/Android)'
  version: 2.0.1
paths:
...
  /api/financeiro:
    get:
      tags:
        - Financeiro
      summary: Listar
      operationId: 3307a46df0686644ad55e8c6b8eb23c2
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
      responses:
        '200':
          description: ''
    post:
      tags:
        - Financeiro
      summary: Cadastrar
      operationId: 9bd840256ac8512f66df62ed7b1c11ea
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
      requestBody:
        description: 'Input data format'
        content:
          multipart/form-data:
            schema: {  }
      responses:
        '201':
          description: ''
  '/api/financeiro/{id}':
    get:
      tags:
        - Financeiro
      summary: Retornar
      operationId: 2bffa669fed333dadea8438c0ac15f4a
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
    put:
      tags:
        - Financeiro
      summary: Atualizar
      operationId: ae16cf1370906abdbeb87c9382fd936c
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: 'Input data format'
        content:
          multipart/form-data:
            schema: {  }
      responses:
        '201':
          description: ''
    delete:
      tags:
        - Financeiro
      summary: Deletar
      operationId: f07d5d0162d3f235ef70dfcd6d42c128
      parameters:
        -
          name: Authorization
          in: header
          required: true
          schema:
            description: 'Bearer token para autorização'
            type: string
            default: 'Bearer <token>'
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
...